Heritage of bank product and utilizate the class mortgage
