+++ src/com/iesam/openbank/domain/BankProducts/Cards.java
@@ -1,0 +1,7 @@
+package com.iesam.openbank.domain.BankProducts;
+
+public class Cards extends BankProduct{
+    public Cards(String code, String name, String numberAccount, String description) {
+        super(code, name, numberAccount, description);
+    }
+}
