+++ src/com/iesam/openbank/domain/BankProducts/FixedTerm.java
@@ -1,0 +1,7 @@
+package com.iesam.openbank.domain.BankProducts;
+
+public class FixedTerm extends BankProduct{
+    public FixedTerm(String code, String name, String numberAccount, String description) {
+        super(code, name, numberAccount, description);
+    }
+}
